pyproject.toml
[project]
name = "deep-agents-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.0.3",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.6",
    "langgraph>=0.6.4",
    "jupyter>=1.1.1",
    "ipykernel>=6.30.1",
    "fastmcp>=2.11.3",
    "tavily-python>=0.7.10",
    "llama-index>=0.12.52",
    "llama-cloud>=0.1.37",
    "wget>=3.2",
    "tqdm>=4.67.1",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-embeddings-ollama>=0.6.0",
    "vllm>=0.8.5.post1",
    "bentoml>=1.4.21",
    "fastapi>=0.116.1",
    "pip-tools>=7.5.0",
    "pyngrok>=7.3.0",
    "llama-index-llms-openai-like>=0.4.0",
    "arize-phoenix-otel>=0.13.0",
    "openinference-instrumentation-llama-index>=4.3.4",
    "arize-phoenix>=11.28.0",
]
